creation interfaces, cleaning code
