+++ src/main/java/fr/postscomments/authentification/register/RegisterController.java
@@ -1,6 +1,7 @@
 package fr.postscomments.authentification.register;
 
 import fr.postscomments.authentification.login.MessageResponse;
+import fr.postscomments.authentification.validationmail.token.ConfirmationTokenService;
 import jakarta.validation.Valid;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -10,24 +11,27 @@ import org.springframework.web.bind.annotation.*;
 public class RegisterController {
 
 
-    private final RegistrationService registrationService;
+    private final RegistrationUserServiceImpl registrationUserServiceImpl;
 
-    public RegisterController(RegistrationService registrationService) {
+    private final ConfirmationTokenService confirmationTokenService;
 
-        this.registrationService = registrationService;
+    public RegisterController(RegistrationUserServiceImpl registrationUserServiceImpl, ConfirmationTokenService confirmationTokenService) {
+
+        this.registrationUserServiceImpl = registrationUserServiceImpl;
+        this.confirmationTokenService = confirmationTokenService;
     }
 
 
     @PostMapping
     public ResponseEntity<MessageResponse> registerUser(@Valid @RequestBody SignUpRequest signUpRequest) {
 
-        String token = registrationService.register(signUpRequest);
+        String token = registrationUserServiceImpl.register(signUpRequest);
         return ResponseEntity.ok(new MessageResponse("user created with success. A message of validation is sended to your adresse mail :" + token));
     }
 
     @GetMapping("/confirm/token={token}")
     public String confirm(@PathVariable("token") String token) {
 
-        return registrationService.confirmToken(token);
+        return confirmationTokenService.confirmToken(token);
     }
 }
