--- src/main/java/fr/postscomments/authentification/register/RegisterUserService.java
@@ -1,6 +1,0 @@
-package fr.postscomments.authentification.register;
-
-public interface RegisterUserService {
-
-
-}
