+++ src/main/java/fr/postscomments/authentification/security/services/UserServices.java
@@ -5,4 +5,10 @@ import fr.postscomments.authentification.models.UserApp;
 public interface UserServices {
 
     UserApp findUserConnected();
+
+    boolean existsByEmail(String email);
+
+    public int enableAppUser(String email);
+
+    UserApp saveUser(UserApp userApp);
 }
