+++ src/main/java/fr/postscomments/authentification/validationmail/email/EmailService.java
@@ -1,0 +1,8 @@
+package fr.postscomments.authentification.validationmail.email;
+
+public interface EmailService {
+
+    public String buildEmail(String name, String link);
+
+    public boolean test(String email);
+}
