--- src/main/java/fr/postscomments/authentification/validationmail/email/EmailValidator.java
@@ -1,15 +1,0 @@
-package fr.postscomments.authentification.validationmail.email;
-
-import org.springframework.stereotype.Service;
-
-import java.util.function.Predicate;
-
-@Service
-public class EmailValidator implements Predicate<String> {
-    private final String EMAIL_PATTERN = "\\w+([-+.']\\w+)*@\\w+([-.]\\w+)*\\.\\w+([-.]\\w+)*";
-
-    @Override
-    public boolean test(String email) {
-        return email.matches(EMAIL_PATTERN);
-    }
-}
