+++ src/main/java/fr/postscomments/authentification/validationmail/token/ConfirmationTokenServiceImpl.java
@@ -1,0 +1,69 @@
+package fr.postscomments.authentification.validationmail.token;
+
+import fr.postscomments.authentification.models.UserApp;
+import fr.postscomments.authentification.security.services.UserServices;
+import jakarta.transaction.Transactional;
+import org.springframework.stereotype.Service;
+
+import java.time.LocalDateTime;
+import java.util.Optional;
+
+@Service
+public class ConfirmationTokenServiceImpl implements ConfirmationTokenService {
+    private final ConfirmationTokenRepository confirmationTokenRepository;
+
+    private final UserServices userServices;
+
+    public ConfirmationTokenServiceImpl(ConfirmationTokenRepository confirmationTokenRepository, UserServices userServices) {
+        this.confirmationTokenRepository = confirmationTokenRepository;
+        this.userServices = userServices;
+    }
+
+    @Override
+    public void saveConfirmationToken(ConfirmationToken token) {
+        confirmationTokenRepository.save(token);
+    }
+
+    @Override
+    public Optional<ConfirmationToken> getToken(String token) {
+        return confirmationTokenRepository.findByToken(token);
+    }
+
+    @Override
+    public int setConfirmedAt(String token) {
+        return confirmationTokenRepository.updateConfirmedAt(token, LocalDateTime.now());
+    }
+
+    @Transactional
+    @Override
+    public String confirmToken(String token) {
+        Optional<ConfirmationToken> confirmToken = getToken(token);
+
+        if (confirmToken.isEmpty()) {
+            throw new IllegalStateException("Token not found!");
+        }
+
+        if (confirmToken.get().getConfirmedAt() != null) {
+            throw new IllegalStateException("Email is already confirmed");
+        }
+
+        LocalDateTime expiresAt = confirmToken.get().getExpiresAt();
+
+        if (expiresAt.isBefore(LocalDateTime.now())) {
+            throw new IllegalStateException("Token is already expired!");
+        }
+
+        setConfirmedAt(token);
+        userServices.enableAppUser(confirmToken.get().getUserApp().getEmail());
+
+        //Returning confirmation message if the token matches
+        return "Your email is confirmed. Thank you for using our service!";
+    }
+
+    @Override
+    public void saveConfirmationToken(UserApp userApp, String token) {
+        ConfirmationToken confirmationToken = new ConfirmationToken(token, LocalDateTime.now(),
+                LocalDateTime.now().plusMinutes(15), userApp);
+        saveConfirmationToken(confirmationToken);
+    }
+}
