+++ src/main/java/fr/postscomments/shared/exceptions/EntityNotValid.java
@@ -1,0 +1,11 @@
+package fr.postscomments.shared.exceptions;
+
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor
+public class EntityNotValid extends RuntimeException {
+    public EntityNotValid(String message) {
+        super(message);
+    }
+
+}
